requirements.txt
asgiref==3.4.0
chardet==4.0.0
Django==3.2.4
djangorestframework==3.12.4
pytz==2021.1
requests==2.25.1
